feat: allow seeding RNG for stable randomness (#16)
* feat: allow seeding RNG for stable randomness

This mostly helps with testing to produce stable images.

The image endpoint accepts an optional `seed` query parameter (`u64`)
which allows passing a seed to the RNG. If absent, the RNG used to
render the image is seeded from the system entropy source.

Also renames the `cause_havoc` query parameter to `wreck_havoc` to
improve the wording.

* feat: show RNG seed in debug logs